commit for piet

+++ mtams/mtams-war/src/java/ManagedBeans/ApprovalBean.java
@@ -215,7 +215,7 @@ public class ApprovalBean implements Serializable {
             appr.setNotes(approvalComment);
             approvalHandler.persistApproval(appr);
             
-            int s = 1;
+            short s = 1;
             selectedAccQte.setSelected(s);
             selectedCarQte.setSelected(s);
             selectedFlgQte.setSelected(s);
